feat: :twisted_rightwards_arrows: Migrate to GTK4 & libadwaita
GTK3 has been superseded by GTK4, most programs in GNOME are now using it.
To better fit in GNOME, the elementary Granite widget should be replaced by libadwaita ones.

💫 Here's a list of what this PR does:
- ⬆️ **Updates** all the deprecated code from GTK3 to GTK4
- 🔳 **Replaces** libgranite widgets with libadwaita
- 🗃️ **Refactors** code for better separation of objects and functions, improving maintainability
- 🐳 **Adds** a devcontainer with some vscode tasks to make it easier to develop

> [!WARNING]
> The translation system is working, but some translations are not longer accurate or functional

> [!NOTE]
> Unless manifested interest from the original developer, I'm not merging this in the original project, some people like 	Granite.
This is a separated port which completely drops GTK3, Granite and ElementaryOS and aims to be as GNOME Circle compliant as possible.

Fix #2
Resolves https://github.com/sdv43/whaler/issues/13
PR #3