chore: :technologist: Move the devcontainer to arch
Moves the devcontainer to archlinux due to a recent bug in Fedora images where containerized Adwaita apps are desaturated on some window manager / DE / distro.

ArchLinux does not seem have this issue and has all the required versions and packages needed to build this project.

If the bug is fixed, this might get reverted.

PR #6